ltiv1p3のaccessToken取得処理を共通化
